// src/renderer/src/api/Album.ts
import { Song } from "./Song";
import { Video } from "./Video";

export interface Album {
  _id: string;
  cover?: string;
  title: string;
  artist: string;
  year?: number;
}

export interface AlbumDetail extends Album {
  genre: string[];
  year: number;
  trackList: Song[];
  videoList: Video[];
}